Data::Tweak is now Data::Overlay
Change prompted by this email response:

On 22/03/11 14:55, Brad Bowman wrote:
> "brian d foy"<brian.d.foy@gmail.com>  wrote:
>>> The following module was proposed for inclusion in the Module List:
>>>
>>> modid:       Data::Tweak DSLIP:       cdphp description: tweak nested
>>> data with composable changes userid:      BOWMANBS (Brad Bowman)
>>
>> It seems taht "nested" and "composable" are more important than "tweak".
>> Should those appear in the name somehow?

While important, "nested" and "composable" are very abstract features
that can apply to lots of scenarios.

> I like "tweak" because it can be the verb (exported sub unlikely to
> collide) and a noun to refer to the change object thing.  It is a bit more
> cutesy than practical though.

"Tweak" is also suggestive of the lightness of shallow copying.

> I'll ponder more searchable options but keep tweak for the development code
> name.

Data::Overlay is the best alternative I've come up with.  It has the same
advantages as "tweak" (except obscurity) and is more suggestive of
non-mutating changes.

I was partly making this submission to get "dibs" on the names,
does it work that way?

Brad