actions on $ds copy, then keys on $new_ds hash
The previous bug was because if $ds was not a HASH, it was
being replaced by an empty hash for keys to override.

I moved the action handling to before the ordinary keys
so that actions can operate on non-HASHes.  If there are
no ordinary keys then $new_ds is returned (can also be not
a HASH).  If there *are* keys in addition to actions, then
$new_ds must be a HASH.